Also rename referrences in python modules
