+++ .regression_files/project/datasets/datasets_test/test_first_batch/fashion_mnist_algorithm_no_op_test.yaml
