+++ .regression_files/project/datasets/datasets_test/test_first_batch/imagenet_algorithm_no_op_validate.yaml
