+++ .regression_files/project/datasets/datasets_test/test_first_batch/mnist_algorithm_no_op_validate.yaml
