+++ project/algorithms/image_classifier_test.py
@@ -7,8 +7,8 @@ from transformers import PreTrainedModel
 from project.algorithms.testsuites.lightning_module_tests import LightningModuleTests
 from project.configs import Config
 from project.conftest import command_line_overrides, skip_on_macOS_in_CI
-from project.datamodules.image_classification.cifar10 import CIFAR10DataModule
-from project.datamodules.image_classification.image_classification import (
+from project.datasets.image_classification.cifar10 import CIFAR10DataModule
+from project.datasets.image_classification.image_classification import (
     ImageClassificationDataModule,
 )
 from project.utils.testutils import run_for_all_configs_of_type
